+++ Plugins/IntelliJ/src/main/kotlin/dev/testify/actions/utility/BaseUtilityAction.kt
@@ -28,15 +28,17 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.OpenFileDescriptor
+import com.intellij.openapi.fileTypes.FileTypeManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiMethod
+import com.intellij.psi.search.FileTypeIndex
 import com.intellij.psi.search.FilenameIndex
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.search.PsiShortNamesCache
+import dev.testify.TestFlavor
 import dev.testify.baselineImageName
 import dev.testify.getVirtualFile
 import org.jetbrains.kotlin.idea.util.projectStructure.module
@@ -46,20 +48,10 @@ abstract class BaseUtilityAction : AnAction() {
 
     override fun getActionUpdateThread() = ActionUpdateThread.EDT
 
-    private fun findClassByName(className: String, project: Project): PsiClass? {
-        val psiShortNamesCache = PsiShortNamesCache.getInstance(project)
-        val classes = psiShortNamesCache.getClassesByName(className, GlobalSearchScope.projectScope(project))
-        return classes.firstOrNull()
-    }
-
     private fun navigateToClass(psiClass: PsiClass, project: Project) {
         val psiFile = psiClass.containingFile.virtualFile
         val descriptor = OpenFileDescriptor(project, psiFile, psiClass.textOffset)
         FileEditorManager.getInstance(project).openTextEditor(descriptor, true)
-    }
-
-    private fun findMethod(methodName: String, psiClass: PsiClass): PsiMethod? {
-        return psiClass.findMethodsByName(methodName, false).firstOrNull()
     }
 
     protected fun navigateToMethod(psiMethod: PsiMethod, project: Project) {
@@ -72,22 +64,39 @@ abstract class BaseUtilityAction : AnAction() {
         val imageFile = this.getVirtualFile()
         val project = this.project
         if (imageFile != null && project != null) {
-            imageFile.nameWithoutExtension.let { imageName ->
-                val parts = imageName.split("_")
-                if (parts.size == 2) {
-                    val (className, methodName) = parts
-                    findClassByName(className, project)?.let { psiClass ->
-                        return findMethod(methodName, psiClass)
-                    }
-                }
+            TestFlavor.entries.forEach { testFlavor ->
+                val method = testFlavor.findSourceMethod(imageFile, project)
+                if (method != null) return method
             }
         }
         return null
     }
 
+    fun findFilesByPartialNameOrRegex(
+        project: Project,
+        partialName: String? = null,
+        regex: Regex? = null,
+        scope: GlobalSearchScope = GlobalSearchScope.projectScope(project)
+    ): List<VirtualFile> {
+        val fileType = FileTypeManager.getInstance().getStdFileType("Image")
+        val allFiles = FileTypeIndex.getFiles(fileType, scope)
+        val fileList = allFiles.filter { file ->
+            when {
+                partialName != null && file.path.contains(partialName, ignoreCase = true) -> true
+                regex != null && regex.matches(file.path) -> true
+                else -> false
+            }
+        }
+
+        return fileList
+    }
+
     protected fun findBaselineImage(currentFile: PsiFile, baselineImageName: String): VirtualFile? {
         if (currentFile is KtFile && currentFile.module != null) {
-            val files = FilenameIndex.getVirtualFilesByName(baselineImageName, currentFile.module!!.moduleContentScope)
+            val files = findFilesByPartialNameOrRegex(
+                project = currentFile.project,
+                partialName = baselineImageName
+            )
             if (files.isNotEmpty()) {
                 return files.first()
             }