+++ Plugins/IntelliJ/src/main/kotlin/dev/testify/actions/utility/DeleteBaselineAction.kt
@@ -29,9 +29,10 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import dev.testify.ConfirmationDialogWrapper
+import dev.testify.TestFlavor
 import java.awt.event.ActionEvent
 
-class DeleteBaselineAction(anchorElement: PsiElement) : BaseFileAction(anchorElement) {
+class DeleteBaselineAction(anchorElement: PsiElement, testFlavor: TestFlavor) : BaseFileAction(anchorElement, testFlavor) {
 
     override val menuText: String
         get() = "Delete ${shortDisplayName(anchorElement)}"